+++ CHANGELOG.md
@@ -27,7 +27,8 @@ development)_
 - Bouncing dots loader tokens file
 - New nativeType prop to the Button component
 - New Checkbox Component
-
+- New ToggleButton and ToggleButtonGroup components
+- Add new icon type for link 
 
 ### Changed